Move some more classes to the "Model" namespaces.
