+++ src/Soloplan.WhatsON.CruiseControl.GUI/CruiseControlTreViewPresentationPlugIn.cs
@@ -11,6 +11,7 @@ namespace Soloplan.WhatsON.CruiseControl.GUI
   using System.IO;
   using System.Text;
   using System.Xml;
+  using Soloplan.WhatsON.CruiseControl.Model;
   using Soloplan.WhatsON.GUI.Common.ConnectorTreeView;
 
   public class CruiseControlTreViewPresentationPlugIn : ITreeViewPresentationPlugIn
