+++ src/Soloplan.WhatsON.CruiseControl/CruiseControlProjectPlugin.cs
@@ -13,6 +13,8 @@ namespace Soloplan.WhatsON.CruiseControl
   using System.Threading.Tasks;
   using Soloplan.WhatsON.Composition;
   using Soloplan.WhatsON.Configuration;
+  using Soloplan.WhatsON.CruiseControl.Model;
+  using Soloplan.WhatsON.Model;
 
   public class CruiseControlProjectPlugin : ConnectorPlugin, IProjectsListQuerying, IAssignServerProject
   {
