+++ src/Soloplan.WhatsON.GUI.Common/ConnectorTreeView/ConnectorGroupViewModel.cs
@@ -13,6 +13,7 @@ namespace Soloplan.WhatsON.GUI.Common.ConnectorTreeView
   using NLog;
   using Soloplan.WhatsON.Composition;
   using Soloplan.WhatsON.Configuration;
+  using Soloplan.WhatsON.Model;
 
   /// <summary>
   /// Viewmodel representing group of connectors shown as single node in <see cref="ConnectorTreeView"/>.
