+++ src/Soloplan.WhatsON.GUI.Common/ConnectorTreeView/ConnectorTreeViewModel.cs
@@ -16,6 +16,7 @@ namespace Soloplan.WhatsON.GUI.Common.ConnectorTreeView
   using NLog;
   using Soloplan.WhatsON.Configuration;
   using Soloplan.WhatsON.GUI.Common.VisualConfig;
+  using Soloplan.WhatsON.Model;
   using Soloplan.WhatsON.Serialization;
 
   /// <summary>
