+++ src/Soloplan.WhatsON.GUI/Configuration/Wizard/WizardController.cs
@@ -19,6 +19,7 @@ namespace Soloplan.WhatsON.GUI.Configuration.Wizard
   using Soloplan.WhatsON.Composition;
   using Soloplan.WhatsON.Configuration;
   using Soloplan.WhatsON.GUI.Configuration.ViewModel;
+  using Soloplan.WhatsON.Model;
   using Soloplan.WhatsON.Serialization;
 
   /// <summary>
