+++ src/Soloplan.WhatsON.GUI/Tests/TrayHandlerTests.cs
@@ -9,6 +9,7 @@ namespace Soloplan.WhatsON.GUI.Tests
   using NUnit.Framework;
   using Soloplan.WhatsON.Configuration;
   using Soloplan.WhatsON.GUI.Common.ConnectorTreeView;
+  using Soloplan.WhatsON.Model;
   using Soloplan.WhatsON.Serialization;
 
   [TestFixture]
