+++ src/Soloplan.WhatsON.Jenkins.GUI/JenkinsProjectViewModel.cs
@@ -11,6 +11,7 @@ namespace Soloplan.WhatsON.Jenkins.GUI
   using Soloplan.WhatsON.GUI.Common;
   using Soloplan.WhatsON.GUI.Common.BuildServer;
   using Soloplan.WhatsON.GUI.Common.ConnectorTreeView;
+  using Soloplan.WhatsON.Model;
 
   public class JenkinsProjectViewModel : BuildServerProjectStatusViewModel
   {
