+++ src/Soloplan.WhatsON.Jenkins.GUI/JenkinsStatusViewModel.cs
@@ -13,6 +13,7 @@ namespace Soloplan.WhatsON.Jenkins.GUI
   using Soloplan.WhatsON.GUI.Common;
   using Soloplan.WhatsON.GUI.Common.BuildServer;
   using Soloplan.WhatsON.Jenkins.Model;
+  using Soloplan.WhatsON.Model;
 
   public class JenkinsStatusViewModel : BuildStatusViewModel
   {
