+++ src/Soloplan.WhatsON.Jenkins/JenkinsProject.cs
@@ -15,6 +15,7 @@ namespace Soloplan.WhatsON.Jenkins
   using Soloplan.WhatsON.Composition;
   using Soloplan.WhatsON.Configuration;
   using Soloplan.WhatsON.Jenkins.Model;
+  using Soloplan.WhatsON.Model;
 
   [ConnectorType("Jenkins Project Status", Description = "Retrieve the current status of a Jenkins project.")]
   [ConfigurationItem(ProjectName, typeof(string), Optional = false, Priority = 300)]
