+++ src/Soloplan.WhatsON.Jenkins/JenkinsProjectPlugin.cs
@@ -14,6 +14,7 @@ namespace Soloplan.WhatsON.Jenkins
   using Soloplan.WhatsON.Composition;
   using Soloplan.WhatsON.Configuration;
   using Soloplan.WhatsON.Jenkins.Model;
+  using Soloplan.WhatsON.Model;
 
   public class JenkinsProjectPlugin : ConnectorPlugin, IProjectsListQuerying, IAssignServerProject
   {
