+++ src/Soloplan.WhatsON/Configuration/ApplicationConfiguration.cs
@@ -9,6 +9,7 @@ namespace Soloplan.WhatsON.Configuration
   using System;
   using System.Collections.Generic;
   using System.Linq;
+  using Soloplan.WhatsON.Model;
 
   /// <summary>
   /// Holds the serializable configuration.
