+++ src/Soloplan.WhatsON/Model/Snapshot.cs
@@ -1,4 +1,4 @@
-﻿namespace Soloplan.WhatsON
+﻿namespace Soloplan.WhatsON.Model
 {
   public class Snapshot
   {
